[refactor] This moves the AbstractConverter class into its own header file. Also add doxygen comments and do some cleanup.
